pyproject.toml
[project]
name = "health-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.7.5",
    "groq>=0.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]
